refactoring the code to right places and added one single test for now
